the beginning of something
It is a really simple disk image creation with a MBR and a "halt loop".

The disk image is created using a linker script. This would be a silly
idea for a "normal" kernel, but it might be an interesting choice for
high-specialised microkernels.

The MBR is compiled with ASM because it needs to. It will be reponsibe
to setup the bare minimal to do jump to "bootstrap" in protected mode.